Update Txs storage to swift async
