+++ unstoppable-ios-app/domains-manager-ios/Services/WalletConnect/WalletConnectExternalWalletHandler/MockWalletConnectExternalWalletHandler.swift
@@ -1,0 +1,8 @@
+//
+//  MockWalletConnectExternalWalletHandler.swift
+//  domains-manager-ios
+//
+//  Created by Oleg Kuplin on 02.03.2023.
+//
+
+import Foundation
