+++ unstoppable-ios-app/domains-manager-ios/Services/WalletConnect/WalletConnectExternalWalletHandler/WalletConnectExternalWalletHandler.swift
