+++ unstoppable-ios-app/domains-manager-ios/Services/WalletConnect/WalletConnectExternalWalletHandler/WalletConnectExternalWalletHandlerProtocol.swift
@@ -1,0 +1,8 @@
+//
+//  WalletConnectExternalWalletHandlerProtocol.swift
+//  domains-manager-ios
+//
+//  Created by Oleg Kuplin on 02.03.2023.
+//
+
+import Foundation
